New Upgrade , New Modified and Merged Sources
Comprehensive migration and refactoring of the Aion game server source code to ensure full compatibility with JDK 25. This commit focuses on removing legacy dependencies, resolving runtime access issues, and modernizing the core engine.
Key Changes:

JDK 25 Migration & Security Patches:

Resolved java.lang.reflect.InaccessibleObjectException by refactoring code segments that required illegal access to internal Java modules.

Replaced deprecated lambdaj library with native Java 8+ Stream API and Collections sorting to ensure high-performance and future-proof compatibility.

Engine Refactoring:

SpawnsData2: Completely refactored to eliminate dependency on lambdaj, fixing NullPointerException and InaccessibleObjectException during server startup.

WalkerFormator & WalkerGroup: Stripped all legacy lambdaj functional calls; implemented modern Java Collectors.groupingBy and List.sort(Comparator) for NPC group organization and pathfinding.

AbyssCore/PvPZones: Resolved critical BufferUnderflowException and instantiation errors by removing legacy mesh-loading dependencies and implementing a streamlined, coordinate-based zone detection system.

Code Cleanup:

Resolved numerous "cannot be resolved to a type" and "unused import" warnings within Eclipse by synchronizing Maven dependencies and cleaning stale compilation artifacts.

Standardized log outputs for critical game services, including DropRegistrationService, to improve debugging transparency.

+++ game-server/src/main/java/com/aionemu/gameserver/model/stats/calc/AdditionStat.java
@@ -39,7 +39,7 @@ public class AdditionStat extends Stat2 {
 
 	@Override
 	public final void addToBonus(int bonus) {
-		this.bonus += (int) (bonusRate * bonus);
+		this.bonus += bonusRate * bonus;
 	}
 
 	@Override
@@ -47,4 +47,5 @@ public class AdditionStat extends Stat2 {
 		return (100 + delta) / 100f;
 	}
 
+	
 }
