+++ game-server/src/main/java/com/aionemu/gameserver/model/stats/calc/ReverseStat.java
@@ -42,13 +42,13 @@ public class ReverseStat extends Stat2 {
 
 	@Override
 	public void addToBonus(int bonus) {
-		this.bonus -= (int) (bonusRate * bonus);
+		this.bonus -= bonusRate * bonus;
 	}
-
+	
 	@Override
 	public float calculatePercent(int delta) {
-		float percent = (100 - delta) / 100f;
-		// TODO need double check here for negatives
+		float percent =  (100 - delta) / 100f;
+		//TODO need double check here for negatives
 		return percent < 0 ? 0 : percent;
 	}
